separate bitcoin
